Subida - clase Martes  7

+++ src/main/java/com/sl/safajobs/servicios/PerfilService.java
@@ -1,11 +1,20 @@
 package com.sl.safajobs.servicios;
 
+import com.sl.safajobs.dto.PerfilCrearDTO;
+import com.sl.safajobs.dto.PerfilDTO;
+import com.sl.safajobs.modelos.Aptitud;
 import com.sl.safajobs.modelos.Perfil;
+import com.sl.safajobs.repositorios.AptitudRepository;
 import com.sl.safajobs.repositorios.PerfilRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Service
 @AllArgsConstructor
@@ -14,14 +23,28 @@ public class PerfilService {
 
     private PerfilRepository perfilRepository;
 
+    private AptitudService aptitudService;
+
 
     /**
      * Este método extrae todos los perfiles de base de datos
      *
      * @return
      */
-    public List<Perfil> getAll(){
-        return perfilRepository.findAll();
+    public List<PerfilDTO> getAll(){
+
+        List<Perfil> perfiles = perfilRepository.findAll();
+        List<PerfilDTO> perfilDTOS = new ArrayList<>();
+
+        for(Perfil p : perfiles){
+            PerfilDTO dto = new PerfilDTO();
+            dto.setNombre(p.getNombre());
+            dto.setApellidos(p.getApellidos());
+            dto.setMail(p.getMail());
+            perfilDTOS.add(dto);
+        }
+
+        return perfilDTOS;
     }
 
 
@@ -39,11 +62,32 @@ public class PerfilService {
     /**
      * Este método guarda un perfil nuevo o modifica uno existente
      *
-     * @param perfil
+     * @param dto
      * @return
      */
-    public Perfil guardar(Perfil perfil){
-        return perfilRepository.save(perfil);
+    public Perfil guardar(PerfilCrearDTO dto){
+
+        Perfil perfilGuardar = new Perfil();
+        perfilGuardar.setNombre(dto.getNombre());
+        perfilGuardar.setApellidos(dto.getApellidos());
+        perfilGuardar.setDni(dto.getDni());
+        perfilGuardar.setMail(dto.getMail());
+
+        //FECHA NACIMIENTO (STRING) -> LOCALTADE
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDate fechaNacimiento = LocalDate.parse(dto.getFechaNacimiento(), formatter);
+        perfilGuardar.setFechaNacimiento(fechaNacimiento);
+
+
+        //APTITUDES
+        Set<Aptitud> aptituds = new HashSet<>();
+        for(Integer idAptitud : dto.getAptitudes()){
+            Aptitud aptitud = aptitudService.getById(idAptitud);
+            aptituds.add(aptitud);
+        }
+        perfilGuardar.setAptitudes(aptituds);
+
+        return perfilRepository.save(perfilGuardar);
     }
 
 
@@ -52,8 +96,27 @@ public class PerfilService {
      *
      * @param id
      */
-    public void eliminar(Integer id){
-        perfilRepository.deleteById(id);
+    public String eliminar(Integer id){
+        String mensaje;
+        Perfil perfil = getById(id);
+
+        if(perfil == null){
+            return  "El perfil con el id indicado no exite";
+        }
+
+        try {
+            perfilRepository.deleteById(id);
+            perfil = getById(id);
+            if(perfil != null){
+                mensaje =  "No se ha podido eliminar el perfil";
+            }else{
+                mensaje = "Perfil eliminado correctamente";
+            }
+        } catch (Exception e) {
+            mensaje =  "No se ha podido eliminar el perfil";
+        }
+
+       return mensaje;
     }
